requirements.txt
spidev
mfrc522
RPi.GPIO
